pyproject.toml
[project]
name = "doc-rendering-mcp-server"
version = "0.1.0"
description = "Document rendering MCP server — templates, HTML/PDF/DOCX/Markdown, Azure Blob Storage delivery"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp[cli]>=1.26.0",
    "httpx>=0.27.0",
    "uvicorn>=0.32.0,<0.34.0",
    "starlette>=0.45.0",
    "jinja2>=3.1.0",
    "weasyprint>=62.0",
    "docxtpl>=0.18.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "azure-storage-blob>=12.23.0",
    "azure-identity>=1.17.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_rendering_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
